+++ test/unit-test/KuksavalUnitTest.cpp
@@ -18,7 +18,7 @@
 #include <string>
 #include <memory>
 #include <limits>
-#include "w3cunittest.hpp"
+#include "KuksavalUnitTest.hpp"
 #include "exception.hpp"
 // #include <jsoncons/json.hpp>
